+++ src/index.ts
@@ -1,6 +1,7 @@
 import { program } from 'commander'
 import pack from '../package.json'
 import { createSignKeyCommand, exportPrivateKey, importPrivateKey, signFileCommand } from './signing'
+import { setSetting } from './app/setting'
 import { getUserId } from './user/account'
 import { zipBundle } from './bundle/zip'
 import { initApp } from './init'
@@ -50,7 +51,6 @@ program
   .action(initApp)
   .option('-n, --name <name>', 'app name')
   .option('-i, --icon <icon>', 'app icon path')
-  .option('-a, --apikey <apikey>', 'apikey to link to your account')
 
 const app = program
   .command('app')
@@ -84,6 +84,13 @@ app
   .option('-a, --apikey <apikey>', 'apikey to link to your account')
   .option('-d, --device <device>', 'the specific device to debug')
   .action(debugApp)
+
+app
+  .command('setting [path]')
+  .description('Modifies capacitor config programmatically')
+  .option('--bool <bool>', 'A value for the setting to modify as a boolean')
+  .option('--string <string>', 'A value for the setting to modify as a string')
+  .action(setSetting)
 
 // app
 //   .command('watch [port]')